Add x264 (MKV) streaming profiles
