+++ pkg/ffmpeg/types.go
@@ -1,6 +1,8 @@
 package ffmpeg
 
 import (
+	"time"
+
 	"github.com/stashapp/stash/pkg/models"
 )
 
@@ -91,4 +93,103 @@ type FFProbeStream struct {
 	MaxBitRate    string `json:"max_bit_rate,omitempty"`
 	SampleFmt     string `json:"sample_fmt,omitempty"`
 	SampleRate    string `json:"sample_rate,omitempty"`
+}
+type Container string
+type AudioCodec string
+
+const (
+	Mp4                Container  = "mp4"
+	M4v                Container  = "m4v"
+	Mov                Container  = "mov"
+	Wmv                Container  = "wmv"
+	Webm               Container  = "webm"
+	Matroska           Container  = "matroska"
+	Avi                Container  = "avi"
+	Flv                Container  = "flv"
+	Mpegts             Container  = "mpegts"
+	Aac                AudioCodec = "aac"
+	Mp3                AudioCodec = "mp3"
+	Opus               AudioCodec = "opus"
+	Vorbis             AudioCodec = "vorbis"
+	MissingUnsupported AudioCodec = ""
+	Mp4Ffmpeg          string     = "mov,mp4,m4a,3gp,3g2,mj2" // browsers support all of them
+	M4vFfmpeg          string     = "mov,mp4,m4a,3gp,3g2,mj2" // so we don't care that ffmpeg
+	MovFfmpeg          string     = "mov,mp4,m4a,3gp,3g2,mj2" // can't differentiate between them
+	WmvFfmpeg          string     = "asf"
+	WebmFfmpeg         string     = "matroska,webm"
+	MatroskaFfmpeg     string     = "matroska,webm"
+	AviFfmpeg          string     = "avi"
+	FlvFfmpeg          string     = "flv"
+	MpegtsFfmpeg       string     = "mpegts"
+	H264               string     = "h264"
+	H265               string     = "h265" // found in rare cases from a faulty encoder
+	Hevc               string     = "hevc"
+	Vp8                string     = "vp8"
+	Vp9                string     = "vp9"
+	MimeWebm           string     = "video/webm"
+	MimeMkv            string     = "video/x-matroska"
+)
+
+var ValidCodecs = []string{H264, H265, Vp8, Vp9}
+
+var validForH264Mkv = []Container{Mp4, Matroska}
+var validForH264 = []Container{Mp4}
+var validForH265Mkv = []Container{Mp4, Matroska}
+var validForH265 = []Container{Mp4}
+var validForVp8 = []Container{Webm}
+var validForVp9Mkv = []Container{Webm, Matroska}
+var validForVp9 = []Container{Webm}
+var validForHevcMkv = []Container{Mp4, Matroska}
+var validForHevc = []Container{Mp4}
+
+var validAudioForMkv = []AudioCodec{Aac, Mp3, Vorbis, Opus}
+var validAudioForWebm = []AudioCodec{Vorbis, Opus}
+var validAudioForMp4 = []AudioCodec{Aac, Mp3}
+
+//maps user readable container strings to ffprobe's format_name
+//on some formats ffprobe can't differentiate
+var ContainerToFfprobe = map[Container]string{
+	Mp4:      Mp4Ffmpeg,
+	M4v:      M4vFfmpeg,
+	Mov:      MovFfmpeg,
+	Wmv:      WmvFfmpeg,
+	Webm:     WebmFfmpeg,
+	Matroska: MatroskaFfmpeg,
+	Avi:      AviFfmpeg,
+	Flv:      FlvFfmpeg,
+	Mpegts:   MpegtsFfmpeg,
+}
+
+var FfprobeToContainer = map[string]Container{
+	Mp4Ffmpeg:      Mp4,
+	WmvFfmpeg:      Wmv,
+	AviFfmpeg:      Avi,
+	FlvFfmpeg:      Flv,
+	MpegtsFfmpeg:   Mpegts,
+	MatroskaFfmpeg: Matroska,
+}
+
+type VideoFile struct {
+	JSON        FFProbeJSON
+	AudioStream *FFProbeStream
+	VideoStream *FFProbeStream
+
+	Path         string
+	Title        string
+	Comment      string
+	Container    string
+	Duration     float64
+	StartTime    float64
+	Bitrate      int64
+	Size         int64
+	CreationTime time.Time
+
+	VideoCodec   string
+	VideoBitrate int64
+	Width        int
+	Height       int
+	FrameRate    float64
+	Rotation     int64
+
+	AudioCodec string
 }
